Fixed a bug and cleaned the code
- fixed a bug in add_edge
- separate file for graph class
- removed inline comments (see styleguide: https://github.com/google/styleguide/blob/gh-pages/pyguide.md#385-block-and-inline-comments)
