+++ group/assignment5/find_alphabet.py
@@ -1,38 +1,6 @@
-"""
-Find the alphabet on the given dictionary + identify inconsistent dictionaries
-"""
-
 from typing import List
-from collections import defaultdict
 from collections import deque
-
-
-class Graph:
-    def __init__(self):
-        self.adjacency_list = defaultdict(list)
-        self.vertices = set()
-
-    def add_edge(self, u, v):
-        self.adjacency_list[u] = v
-        self.vertices.add(u)
-        self.vertices.add(v)
-
-
-def create_graph(dictionary):
-    graph = Graph()
-
-    # go through the dict and compare pairs of words:
-    # one by one compare characters and add to the graph edge f->s and break
-    # where f and s - first non-equal characters from the comparing pair
-    for i in range(len(dictionary)-1):
-        first_word = dictionary[i]
-        second_word = dictionary[i+1]
-        for first_ch, second_ch in zip(first_word, second_word):
-            if first_ch != second_ch:
-                graph.add_edge(first_ch, second_ch)
-                break
-    return graph
-
+from graph import build_graph
 
 def find_alphabet(dictionary: List[str]):
     """
@@ -40,10 +8,7 @@ def find_alphabet(dictionary: List[str]):
     :param dictionary: List of words
     :returns: List of characters or None if the given dictionary is inconsistent
     """
-
-    # create graph from the given dictionary
-    graph = create_graph(dictionary)
-    # return the characters in topological order as the requirement alphabet
+    graph = build_graph(dictionary)
     if graph.vertices:
         result = topological_sort(graph, list(graph.vertices)[0])
         if result is None:
@@ -81,8 +46,6 @@ def topological_sort(graph, v):
     visited = dict.fromkeys(graph.vertices, 0)
     order = deque()
 
-    # go through all vertices and run the updated dfs
-    # if there is a cycle in the given graph => raise Error and return None
     for v in graph.vertices:
         if not visited[v]:
             if not dfs(graph, v, visited, order):
